Middle Commit

+++ app/src/main/java/uz/gita/superdictionary/ui/screen/FavouritesFragment.kt
@@ -60,7 +60,7 @@ class FavouritesFragment : Fragment(R.layout.fragment_favourites) {
         adapter.setItemClick {
             findNavController().navigate(
                 FavouritesFragmentDirections.actionFavouritesFragmentToWordFragment(
-                    it.id.toLong(),
+                    it.id,
                     it.word,
                     it.definition,
                     it.example,
